Online Student Scheduling: Resectioning Weights - fixed an issue with counting same choice, time, room, name criteria   (fractions were computed as integers, always returing 0 or 1)

+++ src/org/cpsolver/studentsct/online/selection/ResectioningWeights.java
@@ -96,7 +96,7 @@ public class ResectioningWeights extends StudentSchedulingAssistantWeights {
                     }
                 }
             }
-            int size = enrollment.getAssignments().size();
+            double size = enrollment.getAssignments().size();
             double sameChoiceFraction = (size - sameChoice) / size;
             double sameTimeFraction = (size - sameTime) / size;
             double sameRoomsFraction = (size - sameRooms) / size;
